pyproject.toml
[build-system]
requires = ["hatchling < 1.22"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["arches_casbin"]

[tool.ruff.lint]
select = ["D", "F", "B"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[project]
name = "arches_casbin"
repository = "https://github.com/flaxandteal/arches_casbin"
readme = "README.md"
version = "0.0.1"
authors = [
  {name = "Phil Weir", email = "phil.weir@flaxandteal.co.uk"},
]
license = "AGPL-3.0-only"
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.11"
dependencies = [
  "casbin-django-orm-adapter",
  "arches_orm @ git+https://github.com/flaxandteal/arches-orm#release/0.2"
]

[tool.hatch.metadata]
allow-direct-references = true

[project.optional-dependencies]
test = [
    'pytest',
    'pytest-asyncio',
    'mock',
    'asynctest'
]
